% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrsa.r
\docType{data}
\name{mff4}
\alias{mff4}
\title{Monthly Fama French 4 Factor Returns}
\format{An \code{xts} object with 1,104 observations and 5 columns. 
The data run from January, 1927 through December, 2018.
As in the upstream source, the data are given in \emph{percents}, meaning a
value of 1.00 corresponds to a 1\% movement.
Note also that returns presumably are \sQuote{simple} returns, not log
returns, though this is not clarified by the upstream source.
The columns are defined as follows:
\describe{
 \item{\code{Mkt}}{The Market monthly return. Note that the risk free rate has been added back to the excess returns published by the upstream source.}
 \item{\code{SMB}}{The cap factor monthly return.}
 \item{\code{HML}}{The growth factor monthly return.}
 \item{\code{UMD}}{The momentum factor monthly return.}
 \item{\code{RF}}{The risk-free rate, presumably as a monthly rate. The average value is arround 0.28, corresponding to an annualized rate of around 3.3\%.}
}}
\source{
Kenneth French data library.
See \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html},
data description at \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/f-f_factors.html}.
}
\usage{
mff4
}
\description{
The monthly returns of the 4 Fama French Factors: Market, the cap factor SMB, 
the growth factor HML, and the momentum factor UMD.
}
\examples{
\dontrun{
data(mff4)
str(mff4)
}
}
\author{
Steven E. Pav \email{steven@sharperat.io}
}
\keyword{datasets}
